+++ src/assets/stylesheets/style.scss
@@ -37,7 +37,7 @@
   }
 
   #background2 {
-    @include background-image('home_bg02.jpg', center, no-repeat, cover);
+    @include background-image('bg_art02.jpg', center, no-repeat, cover);
   }
 
   #background5 {
@@ -98,109 +98,17 @@ main {
   }
 }
 
-/*
-#home-art3 {
-  padding: 0 30px;
-
-  @include flex(row, space-between, center);
-
-  .home-art-sect:nth-child(1) {
-    width: 50%;
-    height: 100%;
-
-    @include background-image('home_art3.png', center, no-repeat, contain);
-  }
-
-  div:nth-child(2) {
-    width: 50%;
-
-    ul {
-      div {
-        @include flex(row, flex-start, center);
-      }
-    }
-  }
-}
-
-#home-art2 {
-  @include flex(column, center, center);
-
-  font-family: 'Amiri', serif;
-  font-weight: 700;
-  //color: $primary-color-medium;
-
-  h3 {
-    font-size: 3rem;
-    margin-bottom: 30px;
-  }
-
-  h4 {
-    font-size: 3.5rem;
-    margin-bottom: 30px;
-  }
-
-  button {
-    padding: 16px 30px 10px 30px;
-    font-family: 'Baloo Tamma 2', cursive;
-    font-weight: 600;
-    font-size: 1.5rem;
-    color: $primary-color-dark;
-    border: none;
-    background-color: $primary-color-light;
-
-    &:hover {
-      color: $primary-color-light;
-      background-color: $primary-color-dark;
-      cursor: pointer;
-    }
-  }
-} */
-
-#art1 {
-  padding-top: 70px;
-
-  @include flex(row, space-between, flex-start);
-
-  text-align: start;
-
-  h2 {
-    margin-bottom: 30px;
-    font-size: 3rem;
-    font-family: 'Amiri', serif;
-  }
-
-  p {
-    font-size: 1.5rem;
-    line-height: 1.9rem;
-  }
-
-  div {
-    padding: 0 60px;
-  }
-
-  div:last-child {
-    h2 {
-      @include background-image('title_bg.jpg', center, repeat, cover);
-
-      -webkit-text-fill-color: transparent;
-      background-clip: content-box;
-      -webkit-background-clip: text;
-      background-attachment: fixed;
-    }
-  }
-}
-
 #art0 {
   @include flex(row, space-between, center);
 
-  div:nth-child(1) {
+  .art0-sect:nth-child(1) {
     width: 50%;
     height: 100%;
 
     @include background-image('home_art0.png', center, no-repeat, contain);
   }
 
-  div:nth-child(2) {
+  .art0-sect:nth-child(2) {
     width: 45%;
     height: 100%;
     padding-top: 10px;
@@ -231,6 +139,96 @@ main {
         background-color: $secondary-color;
         color: $primary-color-dark;
         cursor: pointer;
+      }
+    }
+  }
+}
+
+#art1 {
+  padding-top: 70px;
+
+  @include flex(row, space-between, flex-start);
+
+  text-align: start;
+
+  h2 {
+    margin-bottom: 30px;
+    font-size: 3rem;
+    font-family: 'Amiri', serif;
+  }
+
+  p {
+    font-size: 1.5rem;
+    line-height: 1.9rem;
+  }
+
+  div {
+    padding: 0 60px;
+  }
+
+  .art1-sect:last-child {
+    h2 {
+      @include background-image('title_bg.jpg', center, repeat, cover);
+
+      color: transparent;
+      background-clip: content-box;
+      -webkit-background-clip: text;
+      background-attachment: fixed;
+    }
+  }
+}
+
+#art2 {
+  @include flex(column, center, center);
+
+  font-family: 'Amiri', serif;
+  font-weight: 700;
+
+  h3 {
+    font-size: 3rem;
+    margin-bottom: 30px;
+  }
+
+  h4 {
+    font-size: 3.5rem;
+    margin-bottom: 30px;
+  }
+
+  button {
+    padding: 16px 30px 10px 30px;
+    font-family: 'Baloo Tamma 2', cursive;
+    font-weight: 600;
+    font-size: 1.5rem;
+    color: $primary-color-dark;
+    border: none;
+    background-color: $primary-color-light;
+
+    &:hover {
+      color: $primary-color-light;
+      background-color: $primary-color-dark;
+      cursor: pointer;
+    }
+  }
+}
+
+#art3 {
+  padding: 0 30px;
+
+  @include flex(row, space-between, center);
+
+  .art3-sect:nth-child(1) {
+    width: 50%;
+    height: 100%;
+
+    @include background-image('deals_01.png', center, no-repeat, contain);
+  }
+
+  .art3-sect:nth-child(2) {
+    width: 50%;
+
+    ul {
+      div {
+        @include flex(row, flex-start, center);
       }
     }
   }
